Remove /magic (un-needed) and implement a compilation phase
Previously, the bot would always run in ts-node, regardless of whether
or not it was in dev mode or not. This leads to much higher memory
usage than required, as ts-node *sips* on RAM.

This commit also removes the /magic command that was only needed one
time back in 2022. It is removed, instead of simply disabled, because
it is not possible for bots to control command permissions anymore.
Instead this job belongs to server owners.